pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "csc.teams"
version = "1.0.2"
description = "Team listing addon for Cook Shire Council - displays user teams with LDAP integration"
readme = "README.md"
requires-python = ">=3.11"
authors = [
  {name = "Cook Shire Council", email = "ict@cook.qld.gov.au"}
]
keywords = ["plone", "teams", "ldap", "staff", "directory"]
license = {text = "GPL-2.0-only"}
classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Web Environment",
  "Framework :: Plone",
  "Framework :: Plone :: 6.1",
  "Framework :: Plone :: Addon",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]
dependencies = [
  "Plone>=6.1",
  "plone.api",
  "plone.app.dexterity",
  "plone.app.contenttypes",
  "plone.app.textfield",
  "Products.GenericSetup>=1.8.2",
  "plone.autoform",
  "plone.supermodel",
]

[project.optional-dependencies]
test = [
  "plone.app.testing",
]

[project.urls]
Homepage = "https://github.com/cookshirecouncil/csc.teams"

[tool.setuptools]
package-dir = {"" = "src"}
zip-safe = false
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]
namespaces = false

[tool.setuptools.package-data]
"csc.teams" = [
  "browser/templates/*.pt",
  "browser/static/*.css",
  "browser/static/*.js",
  "profiles/default/*.xml",
  "profiles/default/types/*.xml",
  "profiles/uninstall/*.xml",
]

[project.entry-points."z3c.autoinclude.plugin"]
target = "plone"

[project.entry-points."plone.autoinclude.plugin"]
target = "plone"
